refactor: 도메인 계층 객체 정리
- 문자열 덧셈 계산기 로직 수행에 필요한 Terms(항), DelimiterType(구분자) 객체를 필드로 가지는 Expression 객체 정의
- domain 패키지를 생성하고 Expression, Terms, DelimiterType 객체를 domain 패키지로 이동
