+++ docs/STEP1-README.md
@@ -30,17 +30,6 @@
     * [ ] 구분자 없이 빈값을 입력하는 경우 0 처리
     * [ ] null을 입력하는 경우 예외 처리
 
-* `DelimiterType[enum]` : 구분자 타입 판별 및 추출
-  * [x] 기본 구분자 판별
-    * [x] 콤마
-    * [x] 콜론
-    * [x] 혼합형 : 콤마 + 콜론
-  * [x] 커스텀 구분자 판별
-    * [ ] '//'와 '\n'사이에 위치한 커스텀 구분자 판별 및 구분자 타입인 추출 
-    * [ ] 기본 구분자(콤마, 콜론)이 커스텀 구분자 형식으로 입력된 경우 
-  * 유효성 검증
-    * [x] 기본 구분자 혹은 커스텀 구분자 중 어떤 구분자도 포함되지 않는 경우
-
 * `TermsExtractable[interface]` : 숫자 배열로 이루어진 항(Terms) 목록 추출
   * [x] 판별된 구분자 타입을 기준으로 분리된 연산 대상 목록(List<String>) 추출
     * 기본/커스텀 구분자일 때 항을 추출하는 역할을 수행해야하지만, 그 구현 방식이 다르므로 TermsExtractable `항 추출기`를 추상화 한 후 입력된 문자열의 구분자 타입에 따라 구현체를 선택   
@@ -48,12 +37,25 @@
       * [x] `DefaultTermsExtractor` : 기본 구분자인 경우 구분자 패턴을 기준으로 split()을 이용하여 연산 대상 목록 추출
       * [x] `CustomTermsExtractor` : 커스텀 구분자인 경우 패턴에 일치하는 문자열 그룹 추출 및 split()을 이용한 숫자 문자열 분리
 
-* Terms(class) : 추출된 각 항의 합 연산
-  * [x] 구분자를 기준으로 분리된 List<String>을 입력받아 연산 대상인 숫자 목록으로 변환하여 Terms를 생성하는 정적 팩토리 메소드 작성 
-    * Terms 객체 생성 시 List<String>으로 입력된 각 항에 대한 유효성 검증
-      * [x] 빈값 혹은 null인 원소가 포함된 경우 예외 처리
-      * [x] 숫자가 아니거나 음수인 원소가 포함된 경우 예외 처리
-  * [x] Terms 객체의 각 항을 합산하여 결과를 반환
+* 도메인 객체 설계
+* Expression : 구분자 타입과 연산대상인 항을 필드로 하는 객체
+  * `DelimiterType[enum]` : 구분자 타입 판별 및 추출
+    * [x] 기본 구분자 판별
+      * [x] 콤마
+      * [x] 콜론
+      * [x] 혼합형 : 콤마 + 콜론
+    * [x] 커스텀 구분자 판별
+      * [ ] '//'와 '\n'사이에 위치한 커스텀 구분자 판별 및 구분자 타입인 추출
+      * [ ] 기본 구분자(콤마, 콜론)이 커스텀 구분자 형식으로 입력된 경우
+    * 유효성 검증
+      * [x] 기본 구분자 혹은 커스텀 구분자 중 어떤 구분자도 포함되지 않는 경우 
+  * Terms(class) : 추출된 각 항의 합 연산
+    * [x] 구분자를 기준으로 분리된 List<String>을 입력받아 연산 대상인 숫자 목록으로 변환하여 Terms를 생성하는 정적 팩토리 메소드 작성 
+      * Terms 객체 생성 시 List<String>으로 입력된 각 항에 대한 유효성 검증
+        * [x] 빈값 혹은 null인 원소가 포함된 경우 예외 처리
+        * [x] 숫자가 아니거나 음수인 원소가 포함된 경우 예외 처리
+    * [x] Terms 객체의 각 항을 합산하여 결과를 반환
+
 
 ## 참조 URL
 - [정규표현식의 개념과 기초 문법](https://soooprmx.com/%EC%A0%95%EA%B7%9C%ED%91%9C%ED%98%84%EC%8B%9D%EC%9D%98-%EA%B0%9C%EB%85%90%EA%B3%BC-%EA%B8%B0%EC%B4%88-%EB%AC%B8%EB%B2%95/)
