+++ src/test/kotlin/step1/calculator/domain/DelimiterTypeTest.kt
@@ -1,4 +1,4 @@
-package step1.calculator
+package step1.calculator.domain
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.StringSpec
